business profile and service

+++ config/config.go
@@ -12,10 +12,10 @@ type Config struct {
 }
 
 type AuthConfig struct {
-	MaxFailedLoginAttempts int    `yaml:"max_failed_login_attempts"`
-	LockDurationMinutes    int    `yaml:"lock_duration_minutes"`
-	JWTExpirationHours     int    `yaml:"jwt_expiration_hours"`
-	JWTSecret              string `yaml:"-"`
+	MaxFailedLoginAttempts int     `yaml:"max_failed_login_attempts"`
+	LockDurationMinutes    int     `yaml:"lock_duration_minutes"`
+	JWTExpirationHours     float64 `yaml:"jwt_expiration_hours"`
+	JWTSecret              string  `yaml:"-"`
 }
 
 func Load(path string) (*Config, error) {
